+++ src/rx-error.ts
@@ -6,6 +6,9 @@ import {
     ucfirst
 } from './util';
 import overwritable from './overwritable';
+import {
+    RxErrorParameters
+} from './types';
 
 /**
  * transform an object of parameters to a presentable string
@@ -44,12 +47,12 @@ function messageForError(
 export class RxError extends Error {
     public code: string;
     public message: string;
-    public parameters: any;
+    public parameters: RxErrorParameters;
     public rxdb: true;
     constructor(
         code: string,
         message: string,
-        parameters: any = {}
+        parameters: RxErrorParameters = {}
     ) {
         const mes = messageForError(message, parameters);
         super(mes);
@@ -72,12 +75,12 @@ export class RxError extends Error {
 export class RxTypeError extends TypeError {
     public code: string;
     public message: string;
-    public parameters: any;
+    public parameters: RxErrorParameters;
     public rxdb: true;
     constructor(
         code: string,
         message: string,
-        parameters: any = {}
+        parameters: RxErrorParameters = {}
     ) {
         const mes = messageForError(message, parameters);
         super(mes);
@@ -121,7 +124,7 @@ export function pluginMissing(
 
 export function newRxError(
     code: string,
-    parameters?: any
+    parameters?: RxErrorParameters
 ): RxError {
     return new RxError(
         code,
@@ -131,7 +134,7 @@ export function newRxError(
 }
 export function newRxTypeError(
     code: string,
-    parameters?: any
+    parameters?: RxErrorParameters
 ): RxTypeError {
     return new RxTypeError(
         code,